[Technical] Test for DynamicTableViewController
Note: This doesn't add tests for all methods in this class.
This class is not that easy to test because it is only used as subclasses
and those use storyboards. I've added a fake storyboard to simulate the
loading of those classes in the tests. This is how I would test code like
this. If anyone has a better idea to test this, let me know.