Moved navigationItem init to traitCollectionDidChange()
As requested I moved the initialization of the navigationItem object to the 'traitCollectionDidChange()' method. Please note that I also had to place the initialization  in the custom method 'configureUI()', because the trait method doesn't get called when initializing the app.

I also added a check if the appearance changed, to stop the app from mass recreating the navigationItem when the trait method gets called for another reason.