Exp. Detection error: do not nil out stored summary
When a detection exposure transaction fails due to any error, the cached
summary is cleared from the store - including the timestamp.
This leads to the app continously retrying on the next chance, and
leads to an issue where the user is shown in the UI that risk level can
be re-calculated (in manual mode), although it cannot.