Risk Provider Changes + Fix
Add some early return states when preconditions are not met, such as:
- When tracing has been active for less than 24h (.unknownInitial Risk)
- ExposureManager state is bad (.inactive Risk)

These cases mean that the app will not even attempt to do the async
tasks like exposure detection or downloading the app config unless it is
necessary.

It also fixes an issue where if tracing history was not active for 24h,
the risk calculation would return nil, which now means no risk cell would be shown.